First pass at API Key authorization
This commit allows you to create a specially named secret in a namespace
and allow that secret to be used to authorize requests prior to routing.
To use this, create a generic secret named `ingress` with an `api-key`
data field.  Once you've done this, host and path combinations served by
pods in your namespace will check the value of the `X-INGRESS-API-KEY`
against the value in the secret and return a 403 if they do not match.

See #11