Fix an issue where pod phase change didn't restart
Whenever a pod is added to the internal ingress cache, we need to
trigger an nginx.conf rebuild and an nginx restart.  Prior to this
commit we were only checking if the annotations related to routing
changed but we also need to check the pod's status phase as well.

Fixes #17