Make the router's touch points configurable
Host and path based routing, whether for external (ingress) or internal
purposes is largely the same.  So instead of continuing to tack on more
annotations and configuration items to this router, this commit makes
the touch points configurable so you can use the same code base to run
under multiple different roles.

We also renamed the following configuration/runtime items:

* The `ingress` secret name is now `routing`
* The `publicPaths` annotation is now `routingPaths`
* The `trafficHosts` annotation is now `routingHosts`
* The `X-INGRESS-API-KEY` header is now `X-ROUTING-API-KEY`
* The label selector `microservice=true` now defaults to `routable=true`

See: #11