Always start the router
If you restart a router and there is a pod in a bad state, the initial
restart of nginx after starting with a default configuration will fail.
This means you have to wait until the situation is fixed before the
router can even be deployed.  We should always start the router so that
when the situation is fixed, the router can load the good configuration.

Fixes #37