Validate the routing container port
Prior to this change, you could specify whatever _valid_ port you wanted
in your routing paths and we would accept it.  The problem is if you
specified a container port that was not exposed by a container, we would
blindly accept it and let the routing fail at runtime.  With this change
we will identify this situation and mark the pod as not routable.

Fixes #12